% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustrgame-package.r
\docType{package}
\name{clustrgame-package}
\alias{clustrgame-package}
\title{clustrgame}
\description{
Clustering analysis methods for GPUs.
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\keyword{Package}
