% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.r
\name{km_game}
\alias{km_game}
\title{km}
\usage{
km_game(x, k = 2, maxiter = 100, seed = 1234)
}
\arguments{
\item{x}{A shaq.}

\item{k}{The 'k' in k-means.}

\item{maxiter}{The maximum number of iterations possible.}

\item{seed}{A seed for determining the (random) initial centroids.  Each process has to
use the same seed or very strange things may happen.  If you do not provide
a seed, a good initial seed will be chosen.}
}
\value{
A list containing the cluster centers (global), the observation labels i.e.
the assignments to clusters (distributed shaq), and the total number of
iterations (global).
}
\description{
k-means via Lloyd's Algorithm.
}
\details{
Note that the function does not respect \code{set.seed()} or
\code{comm.set.seed()}.  For managing random seeds, use the \code{seed}
parameter.


The iterations stop either when the maximum number of iterations have been
achieved, or when the centers in the current iteration are basically the same
(within \code{1e-8} for double precision and \code{1e-4} single precision)
as the centers from the previous iteration.

For best performance, the data should be as balanced as possible across all
MPI ranks.
}
\section{Communication}{

Most of the computation is local. However, at each iteration there is a
length \code{n*k} and a length \code{k} allreduce call to update the centers.
There is also a check at the beginning of the call to find out how many
observations come before the current process's data, which is an allgather
operation.
}

\references{
Phillips, J.. Data Mining: Algorithms, Geometry, and Probability.
https://www.cs.utah.edu/~jeffp/DMBook/DM-AGP.html
}
